separate ellipsis in selection bar
