+++ src/features/events/components/SelectionBar.tsx
@@ -5,80 +5,21 @@ import { Button, Divider, Paper, Typography } from '@mui/material';
 import { useSelector, useStore } from 'react-redux';
 
 import messageIds from '../../calendar/l10n/messageIds';
-import { Msg, useMessages } from 'core/i18n';
+import { Msg } from 'core/i18n';
 import { resetSelection } from 'features/events/store';
 import { RootState } from 'core/store';
-import ZUIEllipsisMenu from 'zui/ZUIEllipsisMenu';
-import { useContext } from 'react';
-import { ZUIConfirmDialogContext } from 'zui/ZUIConfirmDialogProvider';
+import SelectionBarEllipsis from './SelectionBarEllipsis';
 
 const SelectionBar = () => {
   const store = useStore<RootState>();
-  const messages = useMessages(messageIds);
-  const { showConfirmDialog } = useContext(ZUIConfirmDialogContext);
 
   const selectedEvents = useSelector(
     (state: RootState) => state.events.selectedEvents
   );
 
-  const events = useSelector(
-    (state: RootState) => state.events.eventList.items
-  );
-
-  const unpublishedEvents = events.filter((event) =>
-    selectedEvents.some(
-      (selectedEvent) =>
-        selectedEvent == event.id && event.data?.published === null
-    )
-  );
-
-  const publishedEvents = events.filter((event) =>
-    selectedEvents.some(
-      (selectedEvent) => selectedEvent == event.id && event.data?.published
-    )
-  );
-
   const handleDeselect = () => {
     store.dispatch(resetSelection());
   };
-
-  const ellipsisMenuItems = [
-    {
-      label: messages.selectionBar.ellipsisMenu.delete(),
-      onSelect: () => {
-        showConfirmDialog({
-          onSubmit: () => {
-            console.log('delete');
-          },
-          title: messages.selectionBar.ellipsisMenu.confirmDelete(),
-          warningText: messages.selectionBar.ellipsisMenu.deleteWarning(),
-        });
-      },
-      textColor: '#ed1c55',
-    },
-    {
-      divider: true,
-      label: messages.selectionBar.ellipsisMenu.cancel(),
-      onSelect: () => {},
-      textColor: '#ed1c55',
-    },
-    {
-      label:
-        publishedEvents.length > 0
-          ? messages.selectionBar.ellipsisMenu.unpublish()
-          : '',
-      onSelect: () => {},
-      textColor: '#f66000',
-    },
-    {
-      label:
-        unpublishedEvents.length > 0
-          ? messages.selectionBar.ellipsisMenu.publish()
-          : '',
-      onSelect: () => {},
-    },
-    { label: messages.selectionBar.ellipsisMenu.print(), onSelect: () => {} },
-  ];
 
   return (
     <Box
@@ -130,7 +71,7 @@ const SelectionBar = () => {
                 <Button endIcon={<ArrowDropDown />} variant="outlined">
                   <Msg id={messageIds.selectionBar.copy} />
                 </Button>
-                <ZUIEllipsisMenu items={ellipsisMenuItems} />
+                <SelectionBarEllipsis />
               </Box>
             </Box>
           </Paper>
